fix utils locations
