+++ optim_esm_tools/plotting/__init__.py
@@ -1,2 +1,3 @@
+from . import plot_utils
 from . import map_maker
 from . import plot
